requirements.txt
argcomplete==3.6.3
asgiref==3.11.1
attrs==26.1.0
cfgv==3.5.0
charset-normalizer==3.4.6
colorama==0.4.6
commitizen==4.13.9
coverage==7.13.5
decli==0.6.3
Deprecated==1.3.1
distlib==0.4.0
Django==6.0.3
django-stubs==6.0.1
django-stubs-ext==6.0.1
djangorestframework==3.17.0
djangorestframework-stubs==3.16.8
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==40.11.0
filelock==3.25.2
identify==2.6.18
inflection==0.5.1
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
librt==0.8.1
MarkupSafe==3.0.3
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
pluggy==1.6.0
pre_commit==4.5.1
prompt_toolkit==3.0.51
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.12.0
python-discovery==1.2.0
PyYAML==6.0.3
questionary==2.1.1
referencing==0.37.0
rpds-py==0.30.0
ruff==0.15.7
sqlparse==0.5.5
termcolor==3.3.0
tomlkit==0.14.0
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
uritemplate==4.2.0
virtualenv==21.2.0
wcwidth==0.6.0
wrapt==2.1.2
